refactor: Improve relative time formatting and MyPage List Layout (#151)
- Introduce `RelativeTimeFormatter.kt` with `toRelativeDayFormat` and `toRelativeTimeFormat` extension functions for `OffsetDateTime` and `LocalDateTime` respectively.
- Update `ProductListItem`, `ChatListScreen`, and `MyPageScreen` to use the new relative time formatting.
- Add new string resources for more granular relative time display (minutes, hours).
- Modify `MyRentalListItem` layout for better visual presentation of dates.
