+++ app/src/main/res/values/strings.xml
@@ -179,8 +179,10 @@
     <string name="screen_product_create_complete_btn_text">등록하기</string>
 
     <!-- 채팅 리스트 -->
-    <string name="screen_chat_list_time_days_ago">· %d일 전</string>
-    <string name="screen_chat_list_time_months_ago">· %d달 전</string>
+    <string name="screen_chat_list_time_minutes_ago">%d분 전</string>
+    <string name="screen_chat_list_time_hours_ago">%d시간 전</string>
+    <string name="screen_chat_list_time_days_ago">%d일 전</string>
+    <string name="screen_chat_list_time_months_ago">%d달 전</string>
     <string name="screen_chat_list_empty_chat_message">아직 채팅이 시작되지 않았어요.</string>
 
     <!-- 채팅방 -->
@@ -229,9 +231,6 @@
     <string name="screen_mypage_tab_title_on_rent">내 대여</string>
     <string name="screen_mypage_text_tab_list_product_empty">게시글이 없어요</string>
     <string name="screen_mypage_text_tab_list_rental_empty">대여 내역이 없어요</string>
-
-    <string name="screen_mypage_my_rental_list_item_label_renting">대여중</string>
-    <string name="screen_mypage_my_rental_list_item_label_request_at">요청일: </string>
 
     <string name="screen_mypage_my_activity_count_label_post">게시글</string>
     <string name="screen_mypage_my_activity_count_label_rental">대여 내역</string>
